Move library modules to subdirectory
